Revert "scripts moved to dir"
This reverts commit 8a0e41812f40fccb3a4aee1e39c349fa118fb855.

+++ code/c/fact.c
@@ -13,7 +13,7 @@ int main(int argc, char const *argv[])
         else
         {
             unsigned long long factorial = 1;
-            for(size_t j = 1; j <= n; j++)
+            for(int j = 1; j <= n; j++)
             {
                 factorial *= j;
             }
